fix: following a pattern on inbox components
